Improves pane sizing code
Instead of calculating the size of the pane manually, by using the padding of the elements put into it, the elements are now placed in a group and this group is then placed in the pane. The pane is then sized according to the group's size, which make the code more flexible, when moving elements around.